requirements.txt
absl-py==2.1.0
amqp==5.2.0
annotated-types==0.6.0
astunparse==1.6.3
beautifulsoup4==4.12.3
billiard==4.2.0
blis==0.7.11
bs4==0.0.2
cachetools==5.3.2
catalogue==2.0.10
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cssselect==1.2.0
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
flatbuffers==23.5.26
flower==2.0.1
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
h5py==3.10.0
humanize==4.9.0
idna==3.6
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
keras==2.15.0
kombu==5.3.5
langcodes==3.3.0
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
ml-dtypes==0.2.0
murmurhash==1.0.10
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
srsly==2.4.8
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.2
tinysegmenter==0.3
tldextract==5.1.1
tornado==6.4
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uuid==1.30
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.14.1
